% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anim.barplot}
\alias{anim.barplot}
\title{Create an animated barplot.}
\usage{
anim.barplot(...)
}
\arguments{
\item{height}{a vector, matrix or array. If a vector it is divided up by
  \code{times} and \code{\link{barplot}} is called on each chunk. If a
  matrix, \code{\link{barplot}} is called on each column. If an array,
  \code{\link{barplot}} is called on each matrix of form \code{height[,,i]}.}
}
\description{
Create an animated barplot.
}
\examples{
anim.barplot(1:100, times=rep(1:10, each=10), ylim=c(0,100))
## barplot with a matrix
ChickWeight$wq <- cut(ChickWeight$weight, 5)
tbl <- as.array(xtabs(~ wq + Diet + Time, data=ChickWeight))
anim.barplot(tbl, xlab="Diet", ylab="N")
anim.barplot(tbl, xlab="Diet", ylab="N", beside=TRUE, legend.text=paste("Quintile", 1:5))
}

