% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anim.plot}
\alias{anim.plot}
\title{Create an animated plot.}
\usage{
anim.plot(x, y, interval = NULL, xlim = NULL, ylim = NULL, col = NULL,
  pch = NULL, cex = NULL, lty = NULL, lwd = NULL, asp = NULL,
  smooth = NULL, ...)
}
\arguments{
\item{x}{a matrix of x values. Each column is plotted in a single frame.}

\item{y}{a matrix of y values.}
}
\description{
\code{anim.plot}
}
\examples{
x <- matrix(rep(-200:200/100, 10), nrow=401, ncol=10)
y <- sin(outer(-200:200/100, 1:10))
anim.plot(x, y, type="l", interval=0.5)
anim.plot(x, y, type="l", interval=0.5, smooth=3)
}

