% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{anim.plots-package}
\alias{anim.plots-package}
\title{Simple Animated Plots For R.}
\description{
\code{anim.plots} contains simple one-line commands to create animated plots
using the \pkg{animation} package. A good place to start is either the
vignette, or the documentation and examples for \code{\link{anim.plot}}.
}

