% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anim.save}
\alias{anim.save}
\title{Save an anim.frames object in various formats.}
\usage{
anim.save(obj, type, filename, smooth.fps = 5, ...)
}
\arguments{
\item{obj}{an \code{anim.frames} object}

\item{type}{one of 'GIF', 'Video', 'SWF', 'HTML', or 'Latex'}

\item{filename}{file to save to}

\item{smooth.fps}{frames per second to plot at}

\item{...}{arguments passed to e.g. \code{\link{saveGIF}}}
}
\description{
This function simply calls replay on the object and then calls
\code{\link{saveGIF}} and friends on the result.
}
\details{
Most underlying functions in the \code{animation} package save at a constant
speed, ignoring timings. To get round this, \code{anim.save} replots
frames every 1/\code{smooth.fps} second.
}
\examples{
\dontrun{
tmp <- anim.plot(1:10, 1:10, pch=1:10, show=FALSE)
anim.save(tmp, "GIF", "filename.gif")

## for anything more complex. Note the curlies:
saveGIF({replay(tmp, after=legend("topleft", legend="My legend"))},
 "filename.gif")
}
}

