% Generated by roxygen2 (4.0.2): do not edit by hand
\name{replay}
\alias{plot.anim.frames}
\alias{replay}
\alias{replay.anim.frames}
\title{Replay an \code{anim.frames} object}
\usage{
replay(...)

\method{replay}{anim.frames}(fr, frames = 1:length(fr), speed = attr(fr,
  "speed"), after = NULL, before = NULL)

\method{plot}{anim.frames}(fr, frames = 1:length(fr), speed = attr(fr,
  "speed"), after = NULL, before = NULL)
}
\arguments{
\item{fr}{an \code{anim.frames} object}

\item{frames}{numeric vector specifying which frames to replay}

\item{speed}{a new speed}

\item{dev.control.enable}{whether to call \code{dev.control('enable')}.
   In most cases the automatic setting will be right.}
}
\description{
Replay all or some of the frames of an object.
}
\examples{
myplot <- anim.plot(1:10, 1:10, speed=3)

replay(myplot, speed=5)
replay(myplot, frames=c(1,5,6,10))
}

