% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveGIF}
\alias{saveGIF}
\title{Save an anim.frames object in various formats.}
\usage{
saveGIF(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{animation::saveGIF}} and such}

\item{obj}{an \code{anim.frames} object}
}
\description{
These functions simply call replay on the object and then call
\code{\link{animation::saveGIF}} and friends on the result.
}
\examples{
\dontrun{
tmp <- anim.plot(1:10, 1:10, pch=1:10, show=FALSE)
saveGIF(tmp, "filename.gif")

## for anything more complex. Note the curlies:
saveGIF({replay(tmp, after=legend("topleft", legend="My legend"))},
 "filename.gif")
}
}

