% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anim.contour}
\alias{anim.contour}
\alias{anim.contour.default}
\alias{anim.filled.contour}
\alias{anim.filled.contour.default}
\title{Create an animated contour plot}
\usage{
anim.contour(...)

anim.filled.contour(...)

\method{anim.filled.contour}{default}(...)

\method{anim.contour}{default}(x, y, z, times, speed = 1, use.times = TRUE,
  window = t, show = TRUE, fn = contour, ...)
}
\arguments{
\item{x,y,z,...}{parameters passed to \code{\link{contour}}}

\item{times,speed,use.times,window,show}{see \code{\link{anim.plot}} for details.}

\item{fn}{underlying contour function to use. For \code{anim.filled.contour}
  this defaults to \code{\link{filled.contour}}.}
}
\description{
Create an animated contour plot
}
\examples{
tmp <- volcano
tmp[] <- 200 - ((row(tmp) - 43)^2 + (col(tmp) - 30)^2)/20
cplot <- array(NA, dim=c(87,61,20))
cplot[,,1] <- tmp
cplot[,,20] <- volcano
cplot <- apply(cplot, 1:2, function(x) seq(x[1], x[20], length.out=20))
cplot <- aperm(cplot, c(2,3,1))
anim.contour(z=cplot, times=1:20, speed=3, levels=80 + 1:12*10, lty=c(1,2,2))
anim.filled.contour(z=cplot, times=1:20, speed=3, levels=80 + 1:12*10,
   color.palette=terrain.colors)
}

