% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anim.contour}
\alias{anim.contour}
\title{Create an animated contour plot}
\usage{
anim.contour(...)
}
\arguments{
\item{x,y,z,...}{parameters passed to \code{\link{contour}}}

\item{times,speed,use.times,window,show}{see \code{\link{anim.plot}} for details.}
}
\description{
Create an animated contour plot
}
\examples{
tmp <- volcano
tmp[] <- 200 - ((row(tmp) - 43)^2 + (col(tmp) - 30)^2)/20
cplot <- array(NA, dim=c(87,61,20))
cplot[,,1] <- tmp
cplot[,,20] <- volcano
cplot <- apply(cplot, 1:2, function(x) seq(x[1], x[20], length.out=20))
cplot <- aperm(cplot, c(2,3,1))
anim.contour(z=cplot, times=1:20, levels=80 + 1:12*10, lty=c(1,2,2))
}

