% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim.plots.R
\docType{package}
\name{anim.plots-package}
\alias{anim.plots}
\alias{anim.plots-package}
\title{anim.plots: simple animated plots For R}
\description{
anim.plots provides simple animated versions of basic R plots, using the 'animation'
package. It includes animated versions of plot, barplot, persp, contour,
filled.contour, hist, curve, points, lines, text, symbols, segments, and
arrows.
}
\details{
For more information, run \code{vignette('anim.plots-stub')}, or check the vignette out
on the web at \url{https://hughjonesd.github.io/anim.plots/anim.plots.html}.


Be aware that anim.plots is just a simple wrapper around Yihui Xie's "animation"
package. You may want to consider more modern solutions such as 
\href{https://cran.r-project.org/package=gganimate}{gganimate}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hughjonesd/anim.plots}
  \item Report bugs at \url{https://github.com/hughjonesd/anim.plots/issues}
}

}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com}

}
