% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anim.segments}
\alias{anim.arrowplot}
\alias{anim.arrows}
\alias{anim.segmentplot}
\alias{anim.segments}
\title{Draw an animation of line segments between points.}
\usage{
anim.segments(x0, y0, x1 = x0, y1 = y0, times, speed = 1, show = TRUE,
  use.times = TRUE, window = t, window.process = NULL, fn = segments,
  col = NULL, lty = NULL, lwd = NULL, ...)

anim.arrows(...)

anim.segmentplot(...)

anim.arrowplot(...)
}
\description{
Draw an animation of line segments between points.
}
\examples{
anim.segments(x0=rep(1:5, 5), y0=rep(1:5, each=5), y1=rep(2:6, each=5),
     times=rep(1:5, each=5) )
}

