% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim.plots.R
\name{PGgame}
\alias{PGgame}
\title{Data from 20 rounds of a public goods game with punishment}
\description{
A 2x3x20 array of data from a laboratory public goods game.
Dimensions are Picked (was subject picked for punishment?),
Contribution (of subject: Non-unique lowest, Not lowest/all same and Unique lowest), and Period.
}
\details{
Provided by the package author.
}
